refactor(machine): add a hook for getting a machine
Move the dispatches for getting a machine to a hook.

Fixes: canonical-web-and-design/app-tribe#1166.
