+++ src/app/machines/components/MachineHeaderForms/ActionFormWrapper/CloneForm/CloneFormFields/CloneFormFields.tsx
@@ -16,6 +16,7 @@ import { actions as machineActions } from "app/store/machine";
 import machineSelectors from "app/store/machine/selectors";
 import type { MachineDetails } from "app/store/machine/types";
 import { isMachineDetails } from "app/store/machine/utils";
+import { useGetMachine } from "app/store/machine/utils/hooks";
 import type { RootState } from "app/store/root/types";
 import { actions as subnetActions } from "app/store/subnet";
 import subnetSelectors from "app/store/subnet/selectors";
@@ -44,6 +45,7 @@ export const CloneFormFields = ({
   const loadingData =
     loadingFabrics || loadingMachines || loadingSubnets || loadingVlans;
   const loadingMachineDetails = !!values.source && !selectedMachine;
+  useGetMachine(values.source);
 
   useEffect(() => {
     dispatch(fabricActions.fetch());
@@ -74,7 +76,6 @@ export const CloneFormFields = ({
         onMachineClick={(machine) => {
           if (machine) {
             setFieldValue("source", machine.system_id);
-            dispatch(machineActions.get(machine.system_id));
           } else {
             setFieldValue("source", "");
             setFieldValue("interfaces", false);
